requirements.txt
flask
flask-cors
numpy
joblib
scikit-learn
opencv-python
mtcnn
Pillow
